refactor: separate Home UI components from project data logic
